Weather Underground API Project
Here's where I'm at so far:
This gets the general location based on the IP address that runs the
code, fetches the current weather, and outputs a forecast for the
current day, the next day, and the day after. I like to start simple and
build onto it.

Process:
1. I played around with the Weather Underground API and used
GetPostMan to quickly analyze the JSON that's received so that I can
figure out what I'd want to put into a database.

I've spent time on the documentation of what's available through the API
located here: https://www.wunderground.com/weather/api/d/docs

2. I decided it'd be fun to grab geolocation data, so I read about how
to do that in Python. I liked to understand what the code I'm
writing/examples I'm following do so that there's few surprises. I opted
to use this:

send_url = 'http://freegeoip.net/json'
r = requests.get(send_url)
j = json.loads(r.text)

Which grabs location based on IP. It's not the most accurate, but it's
simply fetching from a website.

3. I followed the tutorial to learn how to input information in Google
Sheets. I wanted to be able to have it working before I built on that
knowledge.
Once that was set up,  I went to their documentation and read more about
it. The documentation is here:
http://gspread.readthedocs.io/en/latest/#main-interface

As I went, I
would write code in Spyder/Anaconda so that I could continually test it
each time I changed or added something. If it broke, I'd fix it.

One of the challenges was working with JSON and Python, which I had a
little trouble navigating because of the lack of experience with it. It
took me a little time to realize that JSON can be interacted with like
an array. Once I understood that, everything was much easier.

I can grab a list of resources that I used to learn if that's helpful.